Potential function changed to use pow (cmath) and some minor fixes

+++ FunctionsPrototypes.h
@@ -3,6 +3,6 @@
 int Sum(int Num1, int Num2);
 int Subtraction(int Num1, int Num2);
 int Multiplication(int Num1, int Num2);
-int Division(int Num1, int Num2);
+float Division(int Num1, int Num2);
 void Potentiation(int Num1, int Num2);
 void SquareRoot(int Num1, int Num2);
